boton confirmar eliminar xuxemon

+++ xuxemons angular/src/app/components/admin/xuxemons/index/index.component.ts
@@ -24,9 +24,11 @@ export class IndexComponent {
   }
 
   eliminarXuxemon(id: number) {
-    this.xuxemonService.destroy(id).subscribe(
-      () => this.xuxemons = this.xuxemons.filter(xuxemon => xuxemon.id !== id),
-      error => console.error(error)
-    );
+    if (window.confirm('Eliminar Xuxemon?')) {
+      this.xuxemonService.destroy(id).subscribe(
+        () => this.xuxemons = this.xuxemons.filter(xuxemon => xuxemon.id !== id),
+        error => console.error(error)
+      );
+    }
   }
 }
